Eval Refactor Part 1
